modify: migration db tables

+++ school-journal-web/src/main/resources/db/migrate/V1__Create_classes_table.sql
@@ -1,6 +1,7 @@
-CREATE TABLE IF NOT EXISTS `classes`(
-    `id`         int NOT NULL AUTO_INCREMENT,
-    `number` int NOT NULL,
+CREATE TABLE IF NOT EXISTS `classes`
+(
+    `id`        bigint      NOT NULL AUTO_INCREMENT,
+    `number`    int         NOT NULL,
     `character` varchar(32) NOT NULL,
     PRIMARY KEY (`id`)
 ) ENGINE = InnoDB
